taxa fix

+++ src/main/java/com/mycompany/projetopoo/Tarefa.java
@@ -19,7 +19,7 @@ public abstract class Tarefa implements Serializable{
     private GregorianCalendar inicioD;
     private GregorianCalendar finalD;
     private int duracaoEstimada;
-    private double taxa;
+    private int taxa;
     private String nome;
     private Pessoa responsavel;
     private double esforco;
@@ -35,7 +35,7 @@ public abstract class Tarefa implements Serializable{
         setCompleto();
     }
     
-    public Tarefa(String nome,GregorianCalendar inicioD,int duracaoEstimada,Pessoa responsavel,double taxa){
+    public Tarefa(String nome,GregorianCalendar inicioD,int duracaoEstimada,Pessoa responsavel,int taxa){
         this.nome=nome;
         this.inicioD = inicioD;
         this.duracaoEstimada = duracaoEstimada;
@@ -56,7 +56,7 @@ public abstract class Tarefa implements Serializable{
         
     }
     
-    public double getTaxa(){
+    public int getTaxa(){
         return taxa;
     }
